requirements.txt
Django==1.9.5
django-adminplus
celery==3.1.18
redis==2.10.3
requests
django-queryset-csv
MySQL-python
pyvirtualdisplay
selenium==2.45
beautifulsoup4
gevent
grequests
